add command 'all' used to compact both appstate and block state, merge query.go pruning.go and data.go to one
